+++ packages/website/src/pages/Layout.tsx
@@ -418,6 +418,10 @@ function LanguagePicker() {
 			code: "zh",
 			name: "中文",
 		},
+		{
+			code: "sk",
+			name: "Slovak",
+		},
 	]
 
 	const handleSwitchTranslation = (language: { code: string; name: string }) => {