+++ packages/sdk/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @inlang/sdk
 
+## 0.7.0
+
+### Minor Changes
+
+- 0d0502f4: deprecate detectedLanguageTags
+
+### Patch Changes
+
+- Updated dependencies [0d0502f4]
+  - @inlang/plugin@1.3.0
+
 ## 0.6.0
 
 ### Minor Changes